feat(patient): explicit 3NF schema + canonical seed (Story 2.1, NFR-D1/D2)
* New schema.sql declares the 3NF patient table (id BIGINT AUTO_INCREMENT,
  snake_case columns), with DROP+CREATE for deterministic eval-context boot.
* ddl-auto: update -> validate, so entity drift fails fast at startup.
* Patient entity widened Integer -> Long, @Data -> @Getter/@Setter,
  explicit @Column(name="id"); cascade through PatientDTO, PatientRepository,
  PatientService, PatientController.
* data.sql plain INSERT (no IGNORE) of the 4 canonical patients in order,
  yielding ids 1..4 (TestNone, TestBorderline, TestInDanger, TestEarlyOnset)
  for Epic 4 Risk Assessment fixtures.
* New PatientRepositoryTest (@DataJpaTest, Boot 4 packages,
  @AutoConfigureTestDatabase(replace = NONE)) asserts the 4-row seed oracle.
* mvn test -> Tests run: 2, BUILD SUCCESS.

+++ patient-service/src/main/java/com/medilabo/patientservice/controller/PatientController.java
@@ -25,7 +25,7 @@ public class PatientController {
 
     // GET /patients/{id} → détail d'un patient
     @GetMapping("/{id}")
-    public ResponseEntity<PatientDTO> getPatientById(@PathVariable Integer id) {
+    public ResponseEntity<PatientDTO> getPatientById(@PathVariable Long id) {
         return ResponseEntity.ok(patientService.getPatientById(id));
     }
 
@@ -40,7 +40,7 @@ public class PatientController {
     // PUT /patients/{id} → mettre à jour un patient
     @PutMapping("/{id}")
     public ResponseEntity<PatientDTO> updatePatient(
-            @PathVariable Integer id,
+            @PathVariable Long id,
             @Valid @RequestBody PatientDTO dto) {
         return ResponseEntity.ok(patientService.updatePatient(id, dto));
     }
